+++ src/roq/kucoin_futures/live.sh
@@ -16,6 +16,7 @@ FLAGFILE="../../../share/flags/prod/flags.cfg"
 
 MARGIN_MODE="CROSS"
 DOWNLOAD_TRADES_LOOKBACK="24h"
+WS_API=true
 
 $PREFIX ./roq-kucoin-futures \
   --name "$NAME" \
@@ -27,4 +28,5 @@ $PREFIX ./roq-kucoin-futures \
   --service_listen_address "$HOME/run/metrics/${NAME}.sock" \
   --margin_mode="$MARGIN_MODE" \
   --download_trades_lookback="$DOWNLOAD_TRADES_LOOKBACK" \
+  --ws_api=$WS_API \
   $@