+++ src/roq/kucoin_futures/settings.hpp
@@ -20,6 +20,7 @@ struct Settings final : public server::flags::Settings {
   explicit Settings(args::Parser const &);
 
   std::string_view exchange;
+  bool ws_api = {};
   std::string_view margin_mode;
 
   flags::Misc misc;