+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Renamed flag: `--ws_api` (from: `--test_wsapi`)
+
 ## 1.1.1 &ndash; 20251214
 
 ## 1.1.0 &ndash; 2025-11-22